Strip functions out of ImageStacker
